Make db::tables module private
